Add "-mono" option to show monochrome output
One of the features of "nv" was its ability to display video even on
1-bit monochrome displays that could only display pure black and white
without any shades of grey. This was done using a halftoning algorithm.
Since modern X servers typically don't support 1-bit visuals, though,
I needed to add a special switch to force halftoning on, to demonstrate
what it looks like.