Change some defaults
This commit changes the default brightness to 50%. Previously, it
defaulted to 60% to account for older harwdare not accounting for the
gamma of the monitor, but now 50% looks much better.

This commit also changes the max bandwidth limit for 1 Mbps to 10 Mbps,
to allow for the transmit bandwidth slider to support larger transmit
rates.