Update RTPv2 support to match final RFC
This commit updates the RTPv2 support in "nv" to match the final RFCs.
The unreleased 4.0aloha code was based on an earlier Internet Draft
in some places, and the work was never completed.

This version now has basic handling of RTCP sender and receiver reports,
and updates the SDES support to piggyback on the SR/RR packets.